+++ HPXMLtoOpenStudio/tests/test_weather.rb
@@ -187,6 +187,6 @@ class HPXMLtoOpenStudioWeatherTest < Minitest::Test
 
     # Check runner
     assert_equal(0, runner.result.stepErrors.size)
-    assert_equal(1, runner.result.stepWarnings.select { |w| w == 'No design condition info found; calculating design conditions from EPW weather data.' }.size)
+    assert_equal(1, runner.result.stepWarnings.count { |w| w == 'No design condition info found; calculating design conditions from EPW weather data.' })
   end
 end